requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
datasets==3.2.0
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.5
joblib==1.4.2
lightning-utilities==0.11.9
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.1
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
PyAudio==0.2.14
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentry-sdk==2.20.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
SpeechRecognition==3.14.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
torchmetrics==1.6.1
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wandb==0.19.4
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
